commit not working, messing around with GMaps to center on user location.

+++ src/core/app.controller.js
@@ -15,8 +15,6 @@
         vm.coordinates = [];
         vm.title = 'appCtrl';
 
-        // vm.init = GoogleAPIFactory.initMap();
-
         vm.fun_types = ['amusement_park', 'aquarium', 'bar', 'art_gallery', 'book_store', 'campground', 'casino', 'beauty_salon', 'florist', 'movie_rental', 'movie_theater', 'museum', 'night_club', 'park', 'restaurant', 'shopping_mall', 'spa', 'stadium', 'zoo'];
         vm.serious_types = ['airport', 'atm', 'bank', 'car_repair', 'city_hall', 'courthouse', 'dentist', 'doctor', 'electrician', 'embassy', 'fire_station', 'gas_station', 'grocery_or_supermarket', 'gym', 'hardware_store', 'hospital', 'laundry', 'lawyer', 'library', 'locksmith', 'local_government_office', 'lodging', 'painter', 'parking', 'pharmacy', 'plumber', 'police', 'post_office', 'storage', 'train_station', 'transit_station', 'veterinary_care'];
 
@@ -26,8 +24,8 @@
                     var startPos;
                     var geoSuccess = function(position) {
                         startPos = position;
-                        vm.coordinates.push(parseInt(startPos.coords.latitude));
-                        vm.coordinates.push(parseInt(startPos.coords.longitude));
+                        vm.coordinates.push(parseFloat(startPos.coords.latitude));
+                        vm.coordinates.push(parseFloat(startPos.coords.longitude));
                     };
                     navigator.geolocation.getCurrentPosition(geoSuccess);
                 };
